+++ src/components/GroupCreate/ImageSection.tsx
@@ -1,20 +1,31 @@
-import { useState } from 'react';
 import styled from 'styled-components';
 import { IcAddPhoto } from '../../assets';
 
-const ImageSection = () => {
-  const [previewImage, setPreviewImage] = useState<string | null>(null);
+interface ImageSectionProps {
+  previewImage: string | null;
+  handleImageChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  // fileInputKey: number;
+  // handleImageClick: (e: React.MouseEvent<HTMLInputElement>) => void;
+}
 
-  const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { files } = e.target;
-    if (files && files.length > 0) {
-      const reader = new FileReader();
-      reader.onload = () => {
-        setPreviewImage(reader.result as string);
-      };
-      reader.readAsDataURL(files[0]);
-    }
-  };
+const ImageSection = ({
+  previewImage,
+  handleImageChange,
+  // fileInputKey,
+  // handleImageClick,
+}: ImageSectionProps) => {
+  // const [previewImage, setPreviewImage] = useState<string | null>(null);
+
+  // const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+  //   const { files } = e.target;
+  //   if (files && files.length > 0) {
+  //     const reader = new FileReader();
+  //     reader.onload = () => {
+  //       setPreviewImage(reader.result as string);
+  //     };
+  //     reader.readAsDataURL(files[0]);
+  //   }
+  // };
 
   return (
     <Section>
@@ -30,10 +41,12 @@ const ImageSection = () => {
           />
         )}
         <HiddenInput
+          // key={fileInputKey}
           id="fileInput"
           type="file"
           accept="image/*"
           onChange={handleImageChange}
+          // onClick={handleImageClick}
         />
       </ImageContainer>
       <EssentialText>612px * 368px 사이즈를 권장드려요</EssentialText>
@@ -52,6 +65,8 @@ const Label = styled.label`
   gap: 0.6rem;
   align-items: center;
 
+  margin-bottom: 2rem;
+
   ${({ theme }) => theme.fonts.title_bold_20};
   color: ${({ theme }) => theme.colors.white};
 `;
@@ -65,14 +80,6 @@ const ImageContainer = styled.div`
   max-width: 20.4rem;
   cursor: pointer;
 
-  img,
-  svg {
-    max-width: 100%;
-    max-height: 100%;
-
-    border-radius: 0.8rem;
-  }
-
   input {
     display: none;
 
@@ -84,15 +91,13 @@ const HiddenInput = styled.input`
   position: absolute;
   overflow: hidden;
 
-  width: 0;
-  height: 0;
   padding: 0;
 
   border: 0;
 `;
 
 const EssentialText = styled.p`
-  margin-bottom: 1.8rem;
+  margin-top: 0.8rem;
 
   ${({ theme }) => theme.fonts.detail_regular_12};
   color: ${({ theme }) => theme.colors.gray300};
